Refactor and simplification
