--- scripts/safeharbor/active-contracts.csv
@@ -1,3 +1,0 @@
-Status,Chain,Group,Module,Component,Chainlog Key,Address,Etherscan
-ACTIVE,ETHEREUM,VAULTS,Vaults,CDP Manager,CDP_MANAGER,0x5ef30b9986345249bc32d8928B7ee64DE9435E39,https://etherscan.io/address/0x5ef30b9986345249bc32d8928B7ee64DE9435E39#code
-ACTIVE,ETHEREUM,YIELD,Savings,DSR Manager,DSR_MANAGER,0x373238337Bfe1146fb49989fc222523f83081dDb,https://etherscan.io/address/0x373238337Bfe1146fb49989fc222523f83081dDb#code
