+++ scripts/safeharbor/constants.js
@@ -1,0 +1,19 @@
+export const FACTORY_ADDRESS = "0x9d211CaC1ce390F676d1cB1D2Eb681410EC47E47";
+export const OWNER_ADDRESS = "0x195a7d8610edd06e0C27c006b6970319133Cb19A";
+export const AGREEMENT_ADDRESS = "0xa5c82ae35b653192d99c019a8eeaa159af0133e5";
+export const CSV_URL_SHEET1 = "https://docs.google.com/spreadsheets/d/1slHR9VbZOC3wp2ZQu7YbQEQh8N57ePfVvv0w35nz60Q/export?format=csv&gid=1121763694";
+export const CSV_URL_SHEET2 = "https://docs.google.com/spreadsheets/d/1slHR9VbZOC3wp2ZQu7YbQEQh8N57ePfVvv0w35nz60Q/export?format=csv&gid=92820546";
+export const DEFAULT_ASSET_RECOVERY_ADDRESS = "0x0000000000000000000000000000000000000022";
+export const DEFAULT_PROTOCOL_NAME = "Sky and Stars";
+export const DEFAULT_CONTACT = {
+    name: "Contact Name",
+    contact: "Contact@Info"
+};
+export const DEFAULT_BOUNTY_TERMS = {
+    bountyPercentage: 10000000, // 10%
+    bountyCapUSD: 0,
+    retainable: false,
+    identity: 2,
+    diligenceRequirements: "KYC and Sanctions Screening. Sky and Stars require all eligible whitehats to undergo Know Your Customer (KYC) verification and be screened against global sanctions lists, including OFAC, UK, and EU regulations. This ensures that bounty recipients meet legal and regulatory standards before qualifying for payment. The verification process shall be conducted by a trusted third-party provider at Sky and Stars discretion, and all data is deleted, if successful, within 30 days post-verification."
+};
+export const DEFAULT_AGREEMENT_URI = "Agreement URI";
