+++ scripts/safeharbor/package.json
@@ -11,6 +11,7 @@
     },
     "dependencies": {
         "csv-parse": "^5.5.0",
+        "dotenv": "^16.5.0",
         "ethers": "^5.7.2",
         "node-fetch": "^3.3.2"
     },
